chore: separate s3 factory impl
